blockchain-app/test/unit/modules/calt/assets/create_debt_asset.spec.ts
import { CreateDebtAsset } from '../../../../../src/app/modules/calt/assets/create_debt_asset';

describe('CreateDebtAsset', () => {
  let transactionAsset: CreateDebtAsset;

	beforeEach(() => {
		transactionAsset = new CreateDebtAsset();
	});

	describe('constructor', () => {
		it('should have valid id', () => {
			expect(transactionAsset.id).toEqual(4);
		});

		it('should have valid name', () => {
			expect(transactionAsset.name).toEqual('createDebt');
		});

		it('should have valid schema', () => {
			expect(transactionAsset.schema).toMatchSnapshot();
		});
	});

	describe('validate', () => {
		describe('schema validation', () => {
      it.todo('should throw errors for invalid schema');
      it.todo('should be ok for valid schema');
    });
	});

	describe('apply', () => {
    describe('valid cases', () => {
      it.todo('should update the state store');
    });

    describe('invalid cases', () => {
      it.todo('should throw error');
    });
	});
});
